requirements.txt
python-dotenv
streamlit
rich
sentence-transformers
numpy
faiss-cpu

